Avoid dupliced formatters methods via base class
