+++ lib/graphql_connector.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require 'graphql_connector/version'
-require 'graphql_connector/formatters/selected_fields'
+require 'graphql_connector/formatters/base_format'
 require 'graphql_connector/formatters/query_format'
 require 'graphql_connector/formatters/mutation_format'
 require 'graphql_connector/configuration'
